gelf: new private writer type to avoid duplication
The new writer time is responsible of converting creating the message
from a byte slice or taking the specified message as it, and then
converting it into a byte slice which can be passed to an io.Writer.

You MUST specify the encoder options which will be used to create an
encoder which will encode the message into the GELF format before being
passed down to the io.Writer passed in the first argument.