Initial sketch of new architecture
The key functions are `NewStream` and `NewPacket` which take a stream
and packet writer respectively and write new GELF messages to them.
The interface presented to the client is `WriteMessage()` using a
GELFish `Message` struct (allowing full manipulation of almost all
GELF fields) or just a `Write([]byte)` à la io.Writer for when you
don't want to think too much.

The documentation here should be checked with
`godoc -http=127.0.0.1:6060`.